Change MZ_ADLER32_INIT from u64 to u32

+++ miniz_oxide/src/shared.rs
@@ -1,6 +1,6 @@
 use adler32::RollingAdler32;
 
-pub const MZ_ADLER32_INIT : u64 = 1;
+pub const MZ_ADLER32_INIT : u32 = 1;
 
 pub const HUFFMAN_LENGTH_ORDER: [u8; 19] =
     [16, 17, 18, 0, 8, 7, 9, 6, 10, 5, 11, 4, 12, 3, 13, 2, 14, 1, 15];
